+++ tests/external_executor/executors/utils/executor_container.py
@@ -6,11 +6,9 @@ import os
 import threading
 import docker
 import time
-from infra.docker_env import map_workspace_if_azure_devops
 from base64 import b64encode
 from pathlib import Path
 
-from typing import Set, Tuple
 from infra.network import Network
 from infra.node import Node
