+++ tests/external_executor/external_executor.py
@@ -527,20 +527,20 @@ def run(args):
             network = test_wiki_cacher_executor(network, args)
 
     # Run tests with non-containerised initial network
-    # with infra.network.network(
-    #     args.nodes,
-    #     args.binary_dir,
-    #     args.debug_nodes,
-    #     args.perf_nodes,
-    # ) as network:
-    #     network.start_and_open(args)
+    with infra.network.network(
+        args.nodes,
+        args.binary_dir,
+        args.debug_nodes,
+        args.perf_nodes,
+    ) as network:
+        network.start_and_open(args)
 
-    #     network = test_executor_registration(network, args)
-    #     network = test_parallel_executors(network, args)
-    #     network = test_streaming(network, args)
-    #     network = test_async_streaming(network, args)
-    #     network = test_logging_executor(network, args)
-    #     network = test_multiple_executors(network, args)
+        network = test_executor_registration(network, args)
+        network = test_parallel_executors(network, args)
+        network = test_streaming(network, args)
+        network = test_async_streaming(network, args)
+        network = test_logging_executor(network, args)
+        network = test_multiple_executors(network, args)
 
 
 if __name__ == "__main__":
